Applied suggestions
