+++ frontend/src/components/centralErrorAlert/centralErrorAlert.style.ts
@@ -1,0 +1,11 @@
+import type { SxProps } from "@mui/material";
+const errorContainer: SxProps = {
+  display: "flex",
+  justifyContent: "center",
+  alignItems: "center",
+  height: "100%",
+};
+
+const errorAlert: SxProps = { width: "90%" };
+
+export default { errorContainer, errorAlert };
